Move code out of batch_processing and delete unused code
