+++ recordprocessor/src/utils_for_recordprocessor.py
@@ -13,6 +13,13 @@ def get_environment() -> str:
     return _env if _env in ["internal-dev", "int", "ref", "sandbox", "prod"] else "internal-dev"
 
 
+def fetch_file_from_s3(bucket_name, file_key):
+    """Returns the requested file contents in the form of a DictReader"""
+    response = s3_client.get_object(Bucket=bucket_name, Key=file_key)
+    csv_data = response["Body"].read().decode("utf-8")
+    return DictReader(StringIO(csv_data), delimiter="|")
+
+
 def get_csv_content_dict_reader(bucket_name: str, file_key: str):
     """Downloads the csv data and returns a csv_reader with the content of the csv"""
     csv_obj = s3_client.get_object(Bucket=bucket_name, Key=file_key)
